+++ pkgs/num/tests/hash.rs
@@ -8,8 +8,8 @@
 
 #![expect(clippy::unwrap_used, reason = "test code")]
 
-use dash_num::{Hash160, Hash256, ParseHexError};
-use dash_types::Numeric;
+use dash_num::{Hash160, Hash256};
+use dash_types::{Numeric, ParseHexError};
 use hex_literal::hex;
 use rstest::*;